refactored the fields selection into its own component
