+++ frontend/src/components/Layout/FieldSelectMenu/FieldSelectMenu.css
@@ -1,0 +1,3 @@
+.done{
+    left: 75%;
+}
